Dividing the project into files.
