+++ src/main/java/AlgorithmLibrary.java
@@ -1,0 +1,102 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2017-2019 Ilkka Seppälä
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+public class AlgorithmLibrary {
+  /**
+   * Sort an array with quicksort algorithm.
+   *
+   * @param arr   array to sort
+   * @param left  left index where to begin sort (e.g. 0)
+   * @param right right index where to end sort (e.g. array length - 1)
+   */
+  public static void quickSort(int[] arr, int left, int right) {
+    var pivotIndex = left + (right - left) / 2;
+    var pivotValue = arr[pivotIndex];
+    var i = left;
+    var j = right;
+    while (i <= j) {
+      while (arr[i] < pivotValue) {
+        i++;
+      }
+      while (arr[j] > pivotValue) {
+        j--;
+      }
+      if (i <= j) {
+        var tmp = arr[i];
+        arr[i] = arr[j];
+        arr[j] = tmp;
+        i++;
+        j--;
+      }
+      if (left < i) {
+        quickSort(arr, left, j);
+      }
+      if (right > i) {
+        quickSort(arr, i, right);
+      }
+    }
+  }
+
+  /**
+   * Sort an array with bubbleSort algorithm.
+   *
+   * @param arr array to sort
+   */
+  public static void bubbleSort(int[] arr) {
+    var lastIndex = arr.length - 1;
+
+    for (var j = 0; j < lastIndex; j++) {
+      for (var i = 0; i < lastIndex - j; i++) {
+        if (arr[i] > arr[i + 1]) {
+          var tmp = arr[i];
+          arr[i] = arr[i + 1];
+          arr[i + 1] = tmp;
+        }
+      }
+    }
+  }
+
+  /**
+   * Sort an array with selectionSort algorithm.
+   *
+   * @param arr array to sort
+   */
+  public static void selectionSort(int[] arr) {
+    var len = arr.length;
+
+    for (var i = 0; i < len - 1; i++) {
+      var minIndex = i;
+
+      for (var j = i + 1; j < len; j++) {
+        if (arr[j] < arr[minIndex]) {
+          minIndex = j;
+        }
+      }
+
+      var tmp = arr[minIndex];
+      arr[minIndex] = arr[i];
+      arr[i] = tmp;
+    }
+  }
+}
