+++ src/main/java/IoLibrary.java
@@ -1,0 +1,67 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2017-2019 Ilkka Seppälä
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+import java.io.BufferedReader;
+import java.io.FileNotFoundException;
+import java.io.FileReader;
+import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+
+public class IoLibrary {
+  /**
+   * Read file using stream and return list of string lines.
+   *
+   * @param fileName file to read
+   * @throws FileNotFoundException if an I/O error occurs
+   */
+  public static List<String> readFile(String fileName) throws FileNotFoundException {
+    try (Stream<String> stream = new BufferedReader(new FileReader(fileName)).lines()) {
+      return stream.collect(Collectors.toList());
+    }
+  }
+
+  /**
+   * Convert InputStream to String.
+   *
+   * @param inputStream InputStream to convert
+   * @return String
+   * @throws IOException if an I/O error occurs
+   */
+  public static String inputStreamToString(InputStream inputStream) throws IOException {
+    try (var reader = new BufferedReader(new InputStreamReader(inputStream))) {
+      var stringBuilder = new StringBuilder();
+      var data = reader.read();
+
+      while (data != -1) {
+        stringBuilder.append((char) data);
+        data = reader.read();
+      }
+      return stringBuilder.toString();
+    }
+  }
+}
