+++ src/test/java/NetworkingLibraryTest.java
@@ -1,0 +1,60 @@
+/*
+ * MIT License
+ *
+ * Copyright (c) 2017-2019 Ilkka Seppälä
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.containsString;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.io.IOException;
+import java.util.HashMap;
+import org.junit.jupiter.api.Test;
+
+/*
+ * Tests for 30 Seconds of Java code library
+ *
+ */
+class NetworkingLibraryTest {
+  /**
+   * Tests for {@link NetworkingLibrary#httpGet(String)}.
+   */
+  @Test
+  void testHttpGet() throws Exception {
+    var response = NetworkingLibrary.httpGet("http://www.google.com");
+    assertEquals(200, response.statusCode());
+  }
+
+  /**
+   * Tests for {@link NetworkingLibrary#httpPost(String, HashMap)}.
+   */
+  @Test
+  void testHttpPost() throws IOException, InterruptedException {
+    HashMap<String, String> arguments = new HashMap<>();
+    arguments.put("data", "Hello World");
+    var result = NetworkingLibrary.httpPost("https://postman-echo.com/post", arguments);
+    //This postman endpoint echoes the HTTP headers, request parameters, the contents
+    //of the request body and the complete URI requested.
+    var echoedData = "\"data\":\"Hello World\"";
+    assertThat(result.body(), containsString(echoedData));
+  }
+}
